+++ src/components/slider.js
